Configures nexus dogu URL in Dockerfile for better nexus-claim integration
When trying to manually run `nexus-claim` the application server returns
an HTTP 404 not found which is rather empty error message in this
context. As it turns out the default URL is just _almost_ right because
the application server context is moved from / to /nexus/.

This commit introduces a new environment variable within a Docker
container in order to avoid this kind of catch.

The same variable inside the `resources/claim.sh` is left intact,
because this is solely an issue within a docker deployment. Furthermore,
I consider it a good measure of clarity leaving the variable definition
as is.