#37 Support configuration of Nexus log levels
This commit introduces the configurability of Nexus, defaulting to WARN.
Nexus original default, INFO, gives especially the underlying Felix framework
to log an obscene amount of lines which basically can be considered as spam.

Nexus uses 'logback' as log abstraction. The solution in terms of configuration
is to overwrite the whole `logback.xml` file with a template. This follows the
suggested way of implementing a "different" as a simpler configurability does
not seem possible. Please see here:

- https://stackoverflow.com/questions/54487686/how-to-override-logback-config-in-sonatype-nexus-3
- https://community.sonatype.com/t/how-to-override-logback-config-in-nexus-3/1151