#37 Fix log level validations
`doguctl config --default` returns a provided default value in two cases:
- when the queried key exists but consists of the empty string
- when the queried key does not exist at all.

This is in contrast to the template function `.GetOrDefault` which returns a
provided default value **ONLY** if the key does not exists.

This commit changes the validation logic according to above behavior by
accepting defaults only for not existing keys.

Furthermore the validation for the empty string was faulty and needed to be
fixed as well. Here all parameters in question which are likely to contain
empty strings needed to be quoted.