Make task and policy updatable.
This included to make the name of the policy and the type of the task static as these where used to identify a policy/task.
Added configuration values to dogu.json.
Fix various typos.
Update the document to reflect the new changes.

Resolves #108.

Remove orientdb connection credentials from the logs.
This resolves #88.

Update Changelog.